+++ netutil/netutil_darwin_test.go
@@ -35,3 +35,13 @@ func (s *NetUtilSuite) TestGetIP6(c *C) {
 		c.Assert(GetIP6(), Not(Equals), "")
 	}
 }
+
+func (s *NetUtilSuite) TestGetAllIP(c *C) {
+	c.Assert(GetAllIP(), Not(HasLen), 0)
+}
+
+func (s *NetUtilSuite) TestGetAllIP6(c *C) {
+	if os.Getenv("CI") == "" {
+		c.Assert(GetAllIP6(), Not(HasLen), 0)
+	}
+}