+++ netutil/netutil_linux_test.go
@@ -48,6 +48,16 @@ func (s *NetUtilSuite) TestGetIP6(c *C) {
 	}
 }
 
+func (s *NetUtilSuite) TestGetAllIP(c *C) {
+	c.Assert(GetAllIP(), Not(HasLen), 0)
+}
+
+func (s *NetUtilSuite) TestGetAllIP6(c *C) {
+	if os.Getenv("CI") == "" {
+		c.Assert(GetAllIP6(), Not(HasLen), 0)
+	}
+}
+
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 func (s *NetUtilSuite) CreateTestFile(c *C, data string) string {